Bug: 825660 Reviewer: Peter Comment: Fix the case where an ad hoc Filr folder moved into a net folder then deleted can delete the entire net folder

+++ branches/Filr-1.0.0-Bug825660/ssf/main/src/org/kablink/teaming/module/binder/impl/AbstractBinderProcessor.java
@@ -1104,13 +1104,28 @@ public abstract class AbstractBinderProcessor extends CommonDependencyInjection
 							new String[] {binder.getPathName(), driver.getTitle()});
 				}
 				else {
-	    			ResourceSession session = getResourceDriverManager().getSession(driver, ResourceDriverManager.FileOperation.DELETE, binder.getParentBinder()).setPath(binder.getResourcePath());
-	    			try {
-	    				session.delete();
-	    			}
-	    			finally {
-	    				session.close();
-	    			}	
+					//Guard against deleting the whole mirrored source by accident
+					boolean okToDeleteSource = true;
+					if (binder.isAclExternallyControlled() && 
+							binder.getParentBinder() != null && binder.getParentBinder().isAclExternallyControlled() &&
+							binder.getParentBinder().getResourceDriverName().equals(binder.getResourceDriverName())) {
+						
+						//This is a sub-folder of a net folder. Check that it has a proper resource path
+						if (binder.getResourcePath() == null || binder.getResourcePath().equals("") || 
+								binder.getResourcePath().equals("/")) {
+							//Don't allow deleting of this source because it looks like the configuration wasn't properly completed.
+							okToDeleteSource = false;
+						}
+					}
+					if (okToDeleteSource) {
+		    			ResourceSession session = getResourceDriverManager().getSession(driver, ResourceDriverManager.FileOperation.DELETE, binder.getParentBinder()).setPath(binder.getResourcePath());
+		    			try {
+		    				session.delete();
+		    			}
+		    			finally {
+		    				session.close();
+		    			}	
+					}
 				}
     		}
     		catch(NotSupportedException e) {
@@ -1452,6 +1467,9 @@ public abstract class AbstractBinderProcessor extends CommonDependencyInjection
        Binder binder = null;
        try {
 			binder = addBinder(destination, sampleBinder.getEntryDef(), sampleBinder.getClass(), inputData, null, null);
+			//Also copy the configured definitions from the sample
+			binder.setDefinitions(sampleBinder.getDefinitions());
+			getCoreDao().flush();
        } catch (Exception e) {}
        
        return binder;
